# Fix high cpu usage and low throughput