--- a/src/config.rs
+++ b/src/config.rs
@@ -14,7 +14,7 @@ pub struct Config {
 
     /// Max idle connections per worker, trying to create
     /// requests past this limit will fail (return Full?)
-    pub max_idle_connections_per_worker: usize,
+    pub max_connections_per_worker: usize,
 
     /// Max transactions per worker spawned
     pub max_transactions_per_worker: usize,