… .dockerignore
… Cargo.toml
… docker-compose.yaml