When testing against high volume of requests, the current implementation in `master` does poorly.

To fix this:
* Make reading the `Body` of the `Response` optional. 
* Update to hyper 0.12 - this lowered the overall poll count && cpu usage.
* Making DNS resolution in the HttpConnector blocking so redundant resolutions do not occur.

Some benchmarks on my machine (running `time` and sending 2000 requests):

Previous implementation before update (using forked Hyper 0.8):
```
cargo test lots_of_get_single_worker  1.63s user 0.75s system 3% cpu 1:01.42 total
cargo test lots_of_get_single_worker  1.73s user 0.78s system 4% cpu 1:01.28 total
```

Note that it took very little process time. However the total actual time before the responses finish is very slow for some reason.

Current implementation on master:
```
cargo test a_ton_of_gets_single_worker  174.11s user 3.12s system 323% cpu 54.739 total
cargo test a_ton_of_gets_single_worker  150.86s user 2.51s system 337% cpu 45.462 total
cargo test a_ton_of_gets_single_worker  131.68s user 2.09s system 331% cpu 40.413 total
cargo test a_ton_of_gets_single_worker  141.08s user 2.30s system 335% cpu 42.755 total
```

New implementation:
```
cargo test a_ton_of_gets_single_worker -- --nocapture  26.46s user 1.44s system 175% cpu 15.936 total
cargo test a_ton_of_gets_single_worker -- --nocapture  29.30s user 1.52s system 220% cpu 13.986 total
cargo test a_ton_of_gets_single_worker -- --nocapture  20.77s user 1.36s system 97% cpu 22.638 total
cargo test a_ton_of_gets_single_worker -- --nocapture  48.93s user 1.50s system 229% cpu 21.987 total
```

It doesn't use a little process time as the previous implementation, but you can clearly see that it's a huge improvement. The total actual time for the responses to finish is much faster as well.