@jwilm 2018-06-12
Exciting to see progress here!

Synchronous DNS resolution is pretty much a non-starter at the volume of requests we're running. It would block the event loop and not make forward progress for that period.

Regarding higher CPU time, it seems pretty significant. Is it possible that we are polling a big collection or something like that in our implementation? Have you done any profiling here? It would be interesting to see how often our own poll method is called and get an average over number of transactions.