@DarrenTsung 2018-06-12
Will reply with more details later, but good catch about the synchronous
DNS. Actually what we're doing instead of spawning a fully asynchronous
ip_loopup for every request is spawning it on a cpu threadpool that runs
the request synchronously so that it avoids expensive calculations for the
same DNS lookup running in parallel (from the cachegrind report)

Planning to look into profiling it further now that I'm done converting all
the code to use hyper 0.12!

On Mon, Jun 11, 2018, 6:36 PM Joe Wilm <notifications@github.com> wrote:

> Exciting to see progress here!
>
> Synchronous DNS resolution is pretty much a non-starter at the volume of
> requests we're running. It would block the event loop and not make forward
> progress for that period.
>
> Regarding higher CPU time, it seems pretty significant. Is it possible
> that we are polling a big collection or something like that in our
> implementation? Have you done any profiling here? It would be interesting
> to see how often our own poll method is called and get an average over
> number of transactions.
>
> —
> You are receiving this because you were assigned.
> Reply to this email directly, view it on GitHub
> <https://github.com/OneSignal/hyper-client-pool/pull/3#issuecomment-396437938>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABghqQHI5sgcuy90kCkk-CrVjYnz06Akks5t7xs0gaJpZM4UjoqC>
> .
>