@jwilm 2018-06-12
If trust DNS was a problem, let's go back to c-ares. We actually used to get `SIGILL` in production when using hyper's built-in DNS resolution -- we moved away for good reason.