@DarrenTsung 2018-06-12
It doesn't seem to be trust-DNS, using the built-in implementation takes around the same time. Using c-ares actually increases the amount of cpu usage / total time as it seems to fail the requests more often. 

It seems like some of the changes pulled in from updating hyper may be behind the increase of cpu usage. I'm going to run another profiling pass to see if there's anything obvious and check this version integrated with `onepush` to compare the difference.