@DarrenTsung 2018-06-14
After porting this to onepush, I did some more testing on the old commit. See `origin/benchmark-old` for the exact test I used. Turns out that on higher transaction counts hyper starts to break down and start failing more requests and the old commit was ignoring that. I adjusted both tests to run a lower amount of requests (600) so they would mostly complete all the requests successfully.

`benchmark-old` branch:
```
root@c430ddef97aa:/hyper-client-pool# hyperfine --warmup 3 '/hyper-client-pool-target/target/release/deps/hyper_client_pool-51f90383b727a7da lots_of_get'
Benchmark #1: /hyper-client-pool-target/target/release/deps/hyper_client_pool-51f90383b727a7da lots_of_get
  Time (mean ± σ):      6.302 s ±  2.130 s    [User: 401.0 ms, System: 295.0 ms]
  Range (min … max):    5.526 s … 12.346 s

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
```

`cpu-usage` branch:
```
root@c430ddef97aa:/hyper-client-pool# hyperfine --warmup 3 '/hyper-client-pool-target/target/release/deps/lib-f48a956b9a3620cc ton_of_gets'
Benchmark #1: /hyper-client-pool-target/target/release/deps/lib-f48a956b9a3620cc ton_of_gets
  Time (mean ± σ):      2.928 s ±  2.570 s    [User: 863.0 ms, System: 247.0 ms]
  Range (min … max):    1.467 s … 10.097 s
```

Seems like there are still outliers in both test runs, but the user / system usage is much closer! 

Everything works now with `onepush`, but I need to investigate some occasional panics in the dns threads with the new implementation (futures::spawn panicking with [SpawnError::is_shutdown](https://docs.rs/tokio/0.1.7/tokio/executor/struct.SpawnError.html)).