@DarrenTsung 2018-06-15
Okay I spent some time fixing the panicking in trust-dns, but in the end trust-dns was just too slow. I switched back to the c-ares branch and found that it was much more comparable to the previous implementation.

`benchmark-old` results:
```
root@54e5b726bf9a:/hyper-client-pool# hyperfine --warmup 3 '/hyper-client-pool-target/target/release/deps/hyper_client_pool-51f90383b727a7da lots_of_get'
Benchmark #1: /hyper-client-pool-target/target/release/deps/hyper_client_pool-51f90383b727a7da lots_of_get

  Time (mean ± σ):      3.576 s ±  1.739 s    [User: 46.0 ms, System: 145.0 ms]
  Range (min … max):    1.531 s …  5.183 s
```

`hyper-0.12-c-ares`
```
Benchmark #1: /hyper-client-pool-target/target/release/deps/lib-2c3ff674afeb087f ton

  Time (mean ± σ):      4.086 s ±  4.219 s    [User: 80.0 ms, System: 199.0 ms]
  Range (min … max):    1.540 s … 15.158 s

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
```

You can see that c-ares had some outliers, but that was sporadic and seen with the previous implementation as well. 

I think this is good to go! @jwilm @coder543 let me know what you guys think of the latest changes.